requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
sqlmodel==0.0.37
reportlab==4.3.1
python-dotenv==1.0.1
python-multipart==0.0.22
alembic==1.18.4
psycopg2-binary==2.9.11
httpx==0.28.1
